Remove obsolete queue machinery; inline setRootParentSpan
Following the scanner move off effect-app queues (Service Bus / SQLite /
in-memory) onto cluster entities, delete the now-unused queue impls:

- QueueMaker/{SQLQueue,memQueue,sbqueue}.ts: the SB/SQLite/mem QueueMaker
  implementations.
- ServiceBus.ts, memQueue.ts: backing transports for those queues.
- RequestFiberSet.ts: its only remaining consumer was MainFiberSet's
  setRootParentSpan; inline that helper into MainFiberSet directly.

effect-app/QueueMaker.ts keeps only QueueMeta (the per-message
RequestContext envelope still embedded in message schemas); the QueueBase
interface and empty QueueMaker ops object are dropped.
